+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - **_Breaking Change_**: rewrite `IStateMachine` and `StateMachine`, remove `DebugStateMachine`.
 - **_Breaking Change_**: rewrite `IStepExecutor`, `StepExecutor` and `DebugStepExecutor`, remove `InterceptStepExecutor`.
 - **_Breaking Change_**: rewrite `StateManager`.
+- **_Breaking Change_**: remove echoed `Watch` for InputSystem events.
 - Feat: add interface `IReadonlyIoC`.
 - Feat: add `ComponentExtension.GetOrAddComponent`.
 - Feat: add `IReadonlyIoC.GetEventBus`, `IReadonlyIoC.GetCommandBus`, `IReadonlyIoC.GetStepExecutor`, `IIoCC.AddEventBus`, `IIoCC.AddCommandBus` and `IIoCC.AddStepExecutor`.